add JsonHandler for reading/writing JSON file
